nodejs uncluttered

+++ nodejs-expressjs/src/index.ts
@@ -1,40 +1,47 @@
-import express from 'express';
-import { postMiddleware, getMiddleware } from './middleware/middleware.js';
-import bodyParser from 'body-parser';
-import cluster from 'node:cluster';
-import { availableParallelism } from 'node:os';
+import express from "express";
+import { postMiddleware, getMiddleware } from "./middleware/middleware.js";
+import bodyParser from "body-parser";
+import cluster from "cluster";
+import { availableParallelism } from "os";
 
-const numCPUs = Math.max(2, availableParallelism());
-const port = 8080;
+const NUM_CPUS = Math.max(2, availableParallelism());
+const PORT = 8080;
 
 const buildApp = (): express.Express => {
   const app = express();
   app.use(bodyParser.text());
-  app.get('*any', getMiddleware);
-  app.post('*any', postMiddleware);
+  app.get("*any", getMiddleware);
+  app.post("*any", postMiddleware);
   return app;
 };
 
-const gracefulShutdown = (
-  mainProcess: typeof process,
-  sig: 'SIGINT' | 'SIGTERM',
-) => {
+/**
+ * Enable graceful shutdown on a {@link Process}.
+ * @param mainProcess {@link Process}
+ * @param sig {@link Signals}
+ */
+const gracefulShutdown = (mainProcess: NodeJS.Process, sig: NodeJS.Signals) => {
   mainProcess.on(sig, () => {
-    Object.values(cluster.workers).forEach((w) => {
-      w.process.kill(sig);
-    });
+    if (cluster.workers) {
+      Object.values(cluster.workers).forEach((w) => {
+        if (w) w.process.kill(sig);
+      });
+    }
     mainProcess.exit(0);
   });
 };
 
+/**
+ * Cluster and worker launch statements.
+ */
 if (cluster.isPrimary) {
   console.log(`Primary process ${process.pid}`);
-  for (let i = 0; i < numCPUs; i++) {
+  for (let i = 0; i < NUM_CPUS; i++) {
     cluster.fork();
   }
-  process.on('SIGTERM', () => gracefulShutdown(process, 'SIGTERM'));
-  process.on('SIGINT', () => gracefulShutdown(process, 'SIGINT'));
+  process.on("SIGTERM", () => gracefulShutdown(process, "SIGTERM"));
+  process.on("SIGINT", () => gracefulShutdown(process, "SIGINT"));
 } else {
   console.log(`Worker process ${process.pid}`);
-  buildApp().listen(port, '0.0.0.0', () => {});
+  buildApp().listen(PORT, "0.0.0.0", () => {});
 }
